Big change to how keys are handled.
All functions are now accessable thru commands, when a key is pressed
the name string in the key_binding struct is used to lookup the correct
function to call in the cmds[] array. As a result the func field in the
keys[] struct is no longer needed.

This is also a first step towards making it possible to drive a running
xxxterm instance from the command line.

ok marco@