Add a bit of a hack.
If xxxterm crashes at the wrong moment then it has a habit of losing your
session file (this happens to me regularly). So fsync the session file before
closing.

Extending this to do all written files and to save the old one as file~ or
file.backup is something i mean to do soonish.

ok marco