remove gdk_threads_init(); for linux
It hangs the external flash player.  The docs explicitly require this
call however tests show that it isn't required and it makes things
better.  So try using it this way and see if there is some thread
fallout.

I'll take a cluebat as to why this can't be done on linux.