Another day another thread implementation.
The issue comes down to things like flash player are not playing nice
with the gdk mutex.  It does not correctly follow the enter/leave
protocol.  This code now detects this flaw and uses a workaround.  There
is no telling if this code catches all flash player locking violations
but it tries its college best.

Flash is such awesome quality that it often crashes on the way out.  We
can easily ignore that crash so whatever.