add mechanism to run external JavaScript
Add the :script <filename> command to load and execute a piece of
JavaScript in the current tab context.  This should enable all kinds of
awesome extension for xxxterm.  This is also dangerous and not yet
properly cordoned off so use with care.  This feature is experimental.

New things will be added to it as it becomes apparent what else needs to
be done to add things such as password managers etc.