Add autorun JavaScript after page loads
It tries to open a default.js file followed by host.domain.js.  If the
host.domain.js file does not exist it tries to open domain.js.

It reads the content of the default AND host/domain file and executes
it.

All scripts go in the ~/.xxxterm/js/ directory.