Add funky workaround for fancy rich edit boxes in RT
HTML5 defines a new property called "editable".  None of the old
textarea or input properties has that so I never keyed off it.  Turn out
that it is used so treat these as editable things as well.  This works
around an issue in the RT rich text editor which is special in it's own
way.