Parse keybound commands for if they begin with a ':'.  If they do, don't try to execute them, but instead open a prompt with that current command.  Can be used to create custom prompts, and should fix FS#233.
This change also adds the ability to replace strings in keybound
prompts.  So far the only string replacement is <uri> which is
replaced by the current tab's uri.  This also kills the old prompt*
commands as they only existed so they could be bound to a key.
However with the addition of the <uri> substitution, these are no
longer necessary.  Document these changes in the manpage.