This fixes a bunch of things that probably should have gone in my last push.
Make the set_ssl_ca_file pointer logic not suck.

'M' isn't a valid key modifier, use 'M1' in the manpage.

Add a function to escape an html string and use this for displaying
settings.  Now things like <file> and <uri> are not interpreted as
html tags.

Kill a system() when running a custom uri script.  Tildes are now also
expanded instead of being interpreted by the no-longer-being-called
system shell.

Plug a leak

Replace the system() call for external_editor with an execvp and while
here, kill a dangerous sprintf.

Remove an unused variable