Set default char * setting pointers to NULL and g_strdup the default values from static memory, as they will be g_free()'d later if changed at runtime or from parsing the config.  "Special" settings (struct special) do not need this as they use their own set functions.
This also fixes a bad bug where default_script pointed to "" somewhere
in static memory instead of an array of PATH_MAX size.  This fixes a
crash when compiled with clang where changing this setting would try
to write a string possibly as large as PATH_MAX and overwrite other
static data.

Finally, make ssl_ca_file static to match all the other strings that
represent files or directories.  Make it a special setting so static
memory is never freed.  As an added bonus, ssl_ca_file is now tilde
expanded.