Distinguish between toplevel domains and FQDNs in whitelist code.
This fixes the behavior of the whitelist code so that whitelist
toggling toggles the FQDN (not the domain and all subdomains), unless
domain is explicitly used.  This was the intended behavior but the old
code would automatically assume all subdomains anyways.  This also
makes the new force_https stuff work correctly with FQDNs (no
subdomains) for the preloaded HSTS list.