Prevent expanding tabs when using the P keybinding.
This change modifies the paste_uri function to change all newlines in
the uri with spaces.  This stops the new tab from expanding to twice
it's correct height when using the P command, if pasting a link that
contains any newlines (usually these would come from copying the
newline out of a terminal).