Only hide when switching switching tabs
We need to hide t->cmd, t->oops, and t->buffers when creating a new
tab, because when the tab is switched to, they are automatically
hidden in the notebook callback.  Removing this also prevents focusing
the url GtkEntry when using middle click to create a new tab, since
t->focus_wv would not have been set to 1 yet in the create_new_tab()
function.