ref/unref t->active before it is set or unset
This should fix any crashes when determining if t->active is an active
input element or not.  The pointer is from a function which returns
transfer none, so if we want to keep it alive (we do), we need
manually add reference to it, and unrefernce it so it is freed.