Fix back/forward handling with about pages.
This change makes the can_go_* and go_* back/forward functions use the
same logic when determining whether we are on an about page, and makes
it so about:secviolation warning pages do not save the page that
generated the warning to t->item.  This prevents hitting back and
going back to the exact same page that generated the warning
(triggering the warning a second time).